% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR1.r
\name{eda_tunr}
\alias{eda_tunr}
\title{Title eda_tunr : will calculate the following statistics on the datafiles,RMS, Kurtosis, Entropy, Skewness, crestfactor Then it alos calculates the Mahalanobis distance of these features between the different bearings to explore of there is significant correlation}
\usage{
eda_tunr(Result)
}
\arguments{
\item{Result}{is a list of raw vibration datafiles, the sample_size, the number of channels, and the dataset}
}
\value{
a list of datasets for each feature rms,kurtosis, entropy, creatfactor, skewness, and a flatfile will all plus the Mahalanobis   distances
}
\description{
Title eda_tunr : will calculate the following statistics on the datafiles,RMS, Kurtosis, Entropy, Skewness, crestfactor Then it alos calculates the Mahalanobis distance of these features between the different bearings to explore of there is significant correlation
}
\examples{
ans=load_tunr("1st_test",-1)
eda_tunr(ans)

}
