% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR1.r
\name{load_tunr}
\alias{load_tunr}
\title{Title TuneR package to explore vibration datasets
This assumes the NASA data has been extracted to the local directory}
\usage{
load_tunr(type = c("1st_test", "2nd_test"), sample_size)
}
\arguments{
\item{type}{is the list of dataset files to be ingested}

\item{sample_size}{the number of dataset samplles. using 0, or -1 will pick the all files available}
}
\value{
Sets up user specific paths and list of files needed analysis
}
\description{
Title TuneR package to explore vibration datasets
This assumes the NASA data has been extracted to the local directory
}
\examples{
ans=load_tunr("1st_test",10)
ans=load_tunr("1st_test",-1)

}
