% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR1.r
\name{mah}
\alias{mah}
\title{Title Mahalanobis calculation with coarse offset}
\usage{
mah(xx, mah_offset = 0)
}
\arguments{
\item{xx}{this is the dataframe}

\item{mah_offset}{this tries to center the mahanalobis}
}
\value{
single scalar value or vector
}
\description{
Title Mahalanobis calculation with coarse offset
}
\examples{
mah_vec= mah(df)
rms_vec \%<>\% mutate("mah"=mah(rms_vec))
}
