% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneR1.r
\name{plot_tunr}
\alias{plot_tunr}
\title{Title:  plot_tunr will produce plots for the calculated features}
\usage{
plot_tunr(x_type, dataset = "1st_test", regen = FALSE)
}
\arguments{
\item{x_type}{is the freature type to be plotted, options are  "rms", "kurtosis","entropy",",crest","skewness","mahalanobis"}

\item{dataset}{is the name of the dataset "1st_test" or "2nd_test"}

\item{regen}{boolean default is false, only used in the case of re-generate the summary statistics and write out an large tibble}
}
\value{
cool plots
}
\description{
Title:  plot_tunr will produce plots for the calculated features
}
\examples{
ans_df = eda_tunr(load_tunr("1st_test", 20 ))
plot_tunr("Mahalanobis_iid")
plot_tunr("Mahalanobis_ts")

}
